feat(Overview): implement column summaries footer stats with loading and error handling
